Create Exercise_7.1
You are given starting code at the left. The code computes the the square root of the value of variable a. Your task is to:
encapsulate it in a function called mysqrt that takes a as a parameter, chooses a reasonable value of x, and returns an estimate of the square root of a.
To test it, write a function named test_square_root that prints a table like this: